feat: 🎸 client side
