+++ src/utils/pb-payload-helper.ts
@@ -2,7 +2,7 @@ import type {
   puppet,
 } from '@juzi/wechaty-grpc'
 
-import type * as PUPPET from '@juzi/wechaty-puppet'
+import * as PUPPET from '@juzi/wechaty-puppet'
 
 type grpcPuppet = typeof puppet
 
@@ -59,4 +59,55 @@ export const channelPbToPayload = (channelPayloadPb: puppet.ChannelPayload) => {
     ..._channelPayloadPb,
   }
   return channelPayload
+}
+
+export const postPayloadToPb = (grpcPuppet: grpcPuppet, payload: PUPPET.payloads.PostClient) => {
+  const pb = new grpcPuppet.PostPayloadClient()
+  pb.setType(payload.type || 0)
+  for (const item of payload.sayableList) {
+    const sayable = new grpcPuppet.PostSayable()
+    switch (item.type) {
+      case PUPPET.types.Sayable.Text:
+        sayable.setType(grpcPuppet.SayableType.SAYABLE_TYPE_TEXT)
+        sayable.setText(item.payload.text)
+        sayable.setMentionIdListList(item.payload.mentions)
+        break
+      case PUPPET.types.Sayable.Attachment: {
+        sayable.setType(grpcPuppet.SayableType.SAYABLE_TYPE_FILE)
+        const serializedFileBox = typeof item.payload.filebox === 'string' ? item.payload.filebox : await this.serializeFileBox(item.payload.filebox)
+        sayable.setFileBox(serializedFileBox)
+        break
+      }
+      case PUPPET.types.Sayable.Url: {
+        sayable.setType(grpcPuppet.SayableType.SAYABLE_TYPE_URL)
+        const urlLinkPayload = item.payload
+        const pbUrlLinkPayload = urlLinkPayloadToPb(grpcPuppet, urlLinkPayload)
+        sayable.setUrlLink(pbUrlLinkPayload)
+        break
+      }
+      case PUPPET.types.Sayable.Channel: {
+        sayable.setType(grpcPuppet.SayableType.SAYABLE_TYPE_CHANNEL)
+        const channelPayload = item.payload
+        const pbChannelPayload = channelPayloadToPb(grpcPuppet, channelPayload)
+        sayable.setChannel(pbChannelPayload)
+        break
+      }
+      default:
+        throw new Error(`postPublish unsupported type ${item.type}`)
+    }
+    pb.addSayableList(sayable)
+  }
+  if (payload.rootId) { pb.setRootId(payload.rootId) }
+  if (payload.parentId) { pb.setParentId(payload.parentId) }
+  if (payload.location) {
+    const location = new grpcPuppet.LocationPayload()
+    location.setAccuracy(payload.location.accuracy)
+    location.setAddress(payload.location.address)
+    location.setName(payload.location.name)
+    location.setLatitude(payload.location.latitude)
+    location.setLongitude(payload.location.longitude)
+    pb.setLocation(location)
+  }
+  pb.setVisibleListList(payload.visibleList || [])
+  return pb
 }
